Move postajob selenium tests to a management command <PD-1520>
Add user creation method for selenium tests
